Statically compute the query text if possible.
While list placeholders may cause the actual query text
change at runtime, and thus require it to be recalculated
every time the statement is called, it is wasteful to do
so for all queries. When the query parameters do not
contain any lists, it is completely possible to build it
statically during compilation.

Signed-off-by: Alexander Gavrilov <angavrilov@gmail.com>